Clamp night calculations for large world times

+++ src/main/java/com/baddcamden/attributeitemutils/commands/ApplyEnchantsCommand.java
@@ -57,9 +57,9 @@ public class ApplyEnchantsCommand implements CommandExecutor, TabCompleter {
             return true;
         }
 
-        int nights;
+        long nights;
         try {
-            nights = Integer.parseInt(args[1]);
+            nights = Long.parseLong(args[1]);
         } catch (NumberFormatException ex) {
             sender.sendMessage("Invalid nights value: " + args[1]);
             return true;
